% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.group.time.window.R
\name{age.group.time.window}
\alias{age.group.time.window}
\title{Subset the datalist ptable to include only people who existed within the time window and of
agegroup.
Add the eposure start and end time.}
\usage{
age.group.time.window(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(15, 30), site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Individuals within this agegroup}

\item{timewindow}{The window of interest within the simulation time e.g timewindow = c(15, 30)}

\item{site}{Select only the particular site from the study, if "all", use all sites.}
}
\value{
a data.table that only includes people who were alive at the timewindow and are within
the agegroup
}
\description{
Subset the datalist ptable to include only people who existed within the time window and of
agegroup.
Add the eposure start and end time.
}
\examples{
data(datalist)
agegroup.timewindow.df <- age.group.time.window(datalist = datalist, agegroup = c(15, 30),
timewindow = c(10, 40), site="All")

}
