% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agedif.metric.calculator.R
\name{agedif.metric.calculator}
\alias{agedif.metric.calculator}
\title{Calculate different age-mixing metrics, overall and stratified.}
\usage{
agedif.metric.calculator(datalist, agegroup, timepoint, timewindow,
  start = FALSE)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that should be retained, e.g. c(15, 30)}

\item{timepoint}{Point in time at which the age-mixing metrics should be calculated.}

\item{timewindow}{The length of time before the timepoint for which relationships should be included,
e.g. 1, representing one year before the timepoint. This should be a whole number.}

\item{start}{This is a logical indicating that only relationships starting after the beginning of the window
should be used. If start = FALSE relationships could start before the time window.}
}
\value{
a dataframe with median age difference and inter-quartile range(IQR)
for the specified time point and age group, overall, and stratified by gender
}
\description{
Calculate the median age difference at a specified time point,
for the supplied age group, stratfied strata.
}
\examples{
agemix.df <- agedif.metric.calculator(datalist = datalist, agegroup = c(15, 30), timewindow = 1, timepoint = 30)
}

