% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agedistr.creator.R
\name{agedist.create}
\alias{agedist.create}
\title{Create an age distribution for the population at the start of the simulation.}
\usage{
agedist.create(shape = 5, scale = 65)
}
\arguments{
\item{shape}{The shape parameter for the Weibull distribution function (default is 5)}

\item{scale}{The scale parameter for the Weibull distribution function (default is 65)}
}
\value{
a data.frame with Age, Percent.Male and Percent.Female as variables, and 101 rows (0.5 to 100.5 years old). Nobody gets older than 100 years old.
}
\description{
Create an age distribution, consistent with a Weibull survival distribution in the absence of HIV-related mortality, for the population at the start of the simulation.
}
\examples{
agedist.data.frame <- agedist.create(shape = 5, scale = 65)
}

