% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agedif.df.maker.R
\name{agemix.df.maker}
\alias{agemix.df.maker}
\title{Create dataframe to use for age-mixing metrics.}
\usage{
agemix.df.maker(datalist)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}
}
\value{
a dataframe in long format — each row represents an episode in a relationship.
}
\description{
Create a useable dataframe from a simulated population that can then be used for age-mixing metrics.
}
\examples{
agemix.df <- agedif.df.maker(datalist = datalist)
}

