% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemix.episodes.df.maker.R
\name{agemix.episodes.df.maker}
\alias{agemix.episodes.df.maker}
\title{Create dataframe of relationship episodes for agemixing analysis.}
\usage{
agemix.episodes.df.maker(datalist)
}
\arguments{
\item{datalist}{The list object that is produced by \code{\link{readthedata}}}
}
\value{
A dataframe, with the following additional columns generated:
\describe{
    \item{relid}{Unique relationship identier. It combines the unique ID for
    the male parter in the relationship first, and the female partner second.}
    \item{episodeorder}{This indicates the order in which the episodes
    occured.}
    \item{ageepisform}{This indicates the age that the person was when the
    episode was formed.}
    \item{agerelform}{The age of the person when when the relationship
    began. In other words, the age at the first episode.}
    \item{pagerelform}{This is the age of the partner when the
    relationship began.}

 }
}
\description{
Produces a dataframe of relationship episodes from list created by \code{\link{readthedata}}.
}
\details{
The resulting dataframe is in a long format where each row/observation
represents an episode from each relationship. Each relationship episode is
represented twice: once for each man in the dataset who was the male partner
in that relationship, and once for each woman who was the female partner in
the relationship. This allows the user to conduct stratified analyses based
on gender. For relationships where there is more than one episode, there will
be multiple rows.
}
\examples{
cfg <- list()
modeloutput <- RSimpactCyan::simpact.run(configParams = cfg, destDir = "temp")
dl <- readthedata(modeloutput)
agemix.df <- agemix.episodes.df.maker(datalist = dl)

}
