% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemix.rels.df.maker.R
\name{agemix.rels.df.maker}
\alias{agemix.rels.df.maker}
\title{Create dataframe of relationships for agemixing analysis.}
\usage{
agemix.rels.df.maker(dataframe, agegroup, timepoint, timewindow,
  start = FALSE)
}
\arguments{
\item{dataframe}{The dataframe that is produced by \code{\link{agemix.episodes.df.maker}}}

\item{agegroup}{Boundaries of the age group that should be retained, e.g.
c(15, 30). The interval is closed on the left and open on the right.}

\item{timepoint}{Point in time during the simulation to be used in the
calculation.}

\item{timewindow}{The length of time before the timepoint for which
relationships should be included, e.g. 1, representing one year before the
timepoint. This should be a whole number.}

\item{start}{This is a logical indicating that only relationships starting
after the beginning of the window should be used. If start = FALSE
relationships could start before the time window. This is the default.}
}
\value{
returns a dataframe of relationships
}
\description{
Produces a dataframe of relationships from the dataframe of relationship
episodes created by \code{\link{agemix.episodes.df.maker}}. The user
specifies a time point, time window, and age group for which they would like
to obtain this dataframe. This dataframe can then be used as the input for
fitting linear mixed effect models that describe the agemixing pattern with
\code{\link{amp.modeller}}.
}
\examples{
data(persreldf)
agemix.rels.df <- agemix.rels.df.maker(dataframe = persreldf, agegroup = c(15, 30),
timewindow = 1, timepoint = 30, start = FALSE)

}
