% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemixing.2a.wrapper.for.SimpactPaper.R
\name{agemixing.2a.wrapper.for.SimpactPaper}
\alias{agemixing.2a.wrapper.for.SimpactPaper}
\title{Wrapper function for age mixing example of Simpact paper.}
\usage{
agemixing.2a.wrapper.for.SimpactPaper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of model input parameter values}
}
\value{
A vector of model features
}
\description{
This function simulates an HIV epidemic for 20 years and produces model
features for the agemixing pattern, the sexual behaviour, the population
growth rate, HIV prevalence, HIV incidence, and ART coverage.
}
