% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemixing.wrapper.for.SimpactPaper.R
\name{agemixing.wrapper.for.SimpactPaper}
\alias{agemixing.wrapper.for.SimpactPaper}
\title{Wrapper function for age mixing example of Simpact paper.}
\usage{
agemixing.wrapper.for.SimpactPaper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of model input parameter values}
}
\value{
A vector of model features
}
\description{
This function simulates an HIV epidemic for 20 years and produces model
features for the agemixing pattern, the sexual behaviour and the population
growth rate. There is no ART introduced in the model.
}
