% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemixing.wrapper.withhiv.R
\name{agemixing.wrapper.withhiv}
\alias{agemixing.wrapper.withhiv}
\title{Wrapper function to test MICE-assisted calibration functions}
\usage{
agemixing.wrapper.withhiv(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of model input parameter values}
}
\value{
A vector of model features
}
\description{
A short description here...
}
