% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alive.infected.R
\name{alive.infected}
\alias{alive.infected}
\title{Subset people alive at a point in time and adds their HIV status.}
\usage{
alive.infected(datalist = datalist, timepoint = 40, site = "All")
}
\arguments{
\item{datalist}{The list object that is produced by \code{\link{readthedata}}}

\item{timepoint}{Point in time at which the subset should be created and HIV 
status should be evaluated.}

\item{site}{Users can specify a particular site they are interested in. The 
default is "All", which means that persons from all sites should be 
included.}
}
\value{
A data frame that contains a newly generated variable called 
  "Infected" that records the HIV status of persons at the time point
  indicated.
}
\description{
Produces a data frame that contains persons from the simulation that were
alive at a specified point in time. The new data frame also contains a
variable that indicates whether or not the person was infected with HIV at
that point in time.
}
\examples{
cfg <- list()
modeloutput <- RSimpactCyan::simpact.run(configParams = cfg, destDir = "temp")
dl <- readthedata(modeloutput)
alive.at.twenty.df <- alive.infected(datalist = dl, timepoint = 20)

}
