% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.person.years.R
\name{art.person.years}
\alias{art.person.years}
\title{A function that returns the total number of person-years on ART provided between two
points in a simulation time age group and gender}
\usage{
art.person.years(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(15, 40), site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Individuals who have this agegroup within this timewindow}

\item{timewindow}{alive people within this simulation time e.g timewindow = c(15, 30).}

\item{site}{Select only the particular site from the study, if all ignore site/use all sites.}
}
\value{
a total number of people on ART aggregated by Gender.
}
\description{
A function that returns the total number of person-years on ART provided between two
points in a simulation time age group and gender
}
\examples{
data(datalist)
person.years.art <- art.person.years(datalist = datalist, agegroup=c(15, 30),
timewindow =c(15, 30), site="All")
person.years.art

}
