% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.retention.R
\name{ART.retention}
\alias{ART.retention}
\title{A function that returns the percentage of clients who are on treatment
at 6 or more months after ART initiation,
at a paticular point in the simulation time}
\usage{
ART.retention(datalist = datalist, ARTtimewindow = c(20, 34),
  retentiontimeMonths = 12, site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{ARTtimewindow}{the window which treatment was started e.g timepoint = 30.}

\item{retentiontimeMonths}{months from which retention is to be computed e.g 6 or
more months after ART initiation}

\item{site}{Select only the particular site from the study, if all use all sites.}
}
\value{
fraction of people who started treatment and are still on treatment after stated time
}
\description{
Definition. Percentage of adults and children with HIV alive and on ART 12, 24, 36 (etc.)
months after initiating treatment among patients initiating ART during a specified time period.
}
\examples{
data(datalist)
ART.retention <- ART.retention(datalist = datalist, ARTtimewindow= c(30,31),
retentiontimeMonths = 6, site="All")
ART.retention

}
