% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicnumber.calculator.R
\name{basicnumber.calculator}
\alias{basicnumber.calculator}
\title{Basci reproductive number}
\usage{
basicnumber.calculator(datalist = datalist, beta = 0.1508,
  trans.rate.int = trans.rate.int)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{beta}{The probability that an infected individual will infect a suceptible partner over the duration of
their relationship}

\item{trans.rate.int}{A vector of new relationship per time interval produced by \code{\link{transmission.rate.calculator}}}
}
\value{
A value of the basic reproduction number
}
\description{
Compute the  basic reproductive number of a dynamic sexual network simulated with Simpact
}
