% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd.extractor.R
\name{bsd.extractor}
\alias{bsd.extractor}
\title{Compute between-subject standard deviation from agemixing model.}
\usage{
bsd.extractor(models, gender = "male")
}
\arguments{
\item{models}{The models object from which the between-subject standard deviation must
be extracted.}
}
\value{
The between-subject standard deviation
}
\description{
Returns the between-subject standard deviation from the agemixing model that
was fitted with \code{amp.modeller}.
}
\examples{
agemix.episodes.df <- agemix.episodes.df.maker(datalist.agemix)
amp.models <- amp.modeller(dataframe = agemix.episodes.df, agegroup = c(15, 30),
                          timepoint = 15, timewindow = 3, start = FALSE, method = "lme4")

}
