% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd4.atARTinit.calculator.R
\name{cd4.atARTinit.calculator}
\alias{cd4.atARTinit.calculator}
\title{CD4 at ART initiation}
\usage{
cd4.atARTinit.calculator(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(10, 15), cd4count = 350, site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper
bound) that should be retained, e.g. c(15, 30).}

\item{timewindow}{The window in simulation time during which ART initiations are considered.}

\item{cd4count}{The cd4 cell count threshold below which ART initiations are considered.}

\item{site}{Select only the particular site from the study, if "All", use all
sites.}
}
\value{
A dataframe with the number of people who initiated ART below the
  threshold and the total number of people who initiated ART at some point in the time window.
}
\description{
Calculate the number of people who started ART below a given cd4 count
threshold, within a given time window.
}
\examples{
data(datalist)
cd4count.atARTInit <- cd4.atARTinit.calculator(datalist = datalist, agegroup = c(15,50))
cd4count.atARTInit

}
