% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD4.atARTinit.R
\name{cd4.atARTinit}
\alias{cd4.atARTinit}
\title{A function that returns the total number of people who started their treatment when
their CD4count was below a specified threshold segregated by gender}
\usage{
cd4.atARTinit(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(15, 30), cd4count = 350, site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{alive people within this agegroup.}

\item{timewindow}{alive people within this simulation time e.g timewindow = c(20,30).}

\item{cd4count}{threshold at the time of ART initiation.}

\item{site}{select only the particular site from the study, if all ignore site/use all sites.}
}
\value{
the total number of people whose CD4 count at ART initiation is within a was below this threshold
}
\description{
A function that returns the total number of people who started their treatment when
their CD4count was below a specified threshold segregated by gender
}
\examples{
data(datalist)
cd4.atARTinit <- cd4.atARTinit(datalist = datalist, agegroup=c(15, 30),
 timewindow=c(10, 40), cd4count=350, site="All")

}
