% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.sequence.ind.R
\name{choose.sequence.ind}
\alias{choose.sequence.ind}
\title{Subset sequences in fasta file based on sequence ID or Name}
\usage{
choose.sequence.ind(pool.seq.file = "pool.seq.fasta",
  select.vec = select.vec, name.file = "selected.ind.fasta")
}
\arguments{
\item{pool.seq.file}{File of sequence data (in fasta format) within which we want to subset}

\item{select.vec}{Vector of ID or name of sequences we want to select}

\item{name.file}{Name of fasta file which containes selected sequences}
}
\value{
Files of elcted sequences data
}
\description{
Subset sequences in fasta file based on sequence ID or Name
}
