% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent.point.prevalence.R
\name{concurr.pointprev.calculator}
\alias{concurr.pointprev.calculator}
\title{Calculate concurrent point prevalence, overall and stratified.}
\usage{
concurr.pointprev.calculator(datalist = datalist, agegroup = c(15, 50),
  timepoint, hivstatus = 2)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that should
be retained, e.g. c(15, 30)}

\item{timepoint}{Point in time at which the HIV prevalence should be calculated.}

\item{hivstatus}{HIV status at the time of the survey. Options are 2, means all; 0
means only HIV-negative, 1 means only HIV-positive.}
}
\value{
A dataframe with concurrent point prevalence estimate and surrounding confidence bounds,
for the specified time point, age group and hiv status, overall, and stratified by gender
}
\description{
Calculate the concurrent point prevalence, for specific age groups and gender strata.
}
\examples{
data(datalist)
concurrent.pointprev.df <- concurr.pointprev.calculator(datalist = datalist, agegroup = c(15,50),
timepoint = 10, hivstatus = 2)
concurrent.pointprev.df

}
