% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.person.df.maker.R
\name{coxph.person.df.maker}
\alias{coxph.person.df.maker}
\title{Create a dataframe for Cox Proportional Hazard Regression analysis at the person-level.}
\usage{
coxph.person.df.maker(datalist = datalist.test, agegroup = c(15, 30),
  timewindow = c(30, 40))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that should be retained, e.g. c(15, 30)}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound) that should be retained, e.g. c(30, 40)}
}
\value{
a dataframe of individuals, with age at the time of entering the cohort (age.start),
age at the time of event or censoring (i.e. reaching agegroup boundary or time window boundary) time (age.end),
and an indicator of whether HIV infection happened.
}
\description{
Create a dataframe with one row per woman who is at least for some time during the time window HIV-negative and between 15 and 30 years old.
}
\details{
We can treat the data as left-truncated (women must be HIV-negative at the time of entering the cohort of 15>= year old women).
Because we assume that the study cohort's observation time starts at some arbitrary point in the simulation
(e.g. 20 years after HIV is introduced), some women will enter the cohort at an age > 15 (but still < 30).
time is the left-truncated age of entering the cohort
time2 is the age at event time or age at the end of the time window or agegroup if the event wasn't observed by then.
The event parameter indicates whether or not the HIV transmission event was observed or not.
type indicates that we have right-censored data.
}
\examples{
coxph.person.df <- coxph.person.df.maker(datalist = datalist, agegroup = c(15, 30), timewindow = c(30, 40))
}

