% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum.network.R
\name{cum.network}
\alias{cum.network}
\title{Build a cumulative relationships network - an igraph object}
\usage{
cum.network(datalist = datalist, endtime = 15)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{endtime}{Only relationships events that took place before this point in simulation time, are captured in the output.}
}
\value{
A graph object for relationships network data with difference between gender and HIV status
Men are represented by square, women by sphere, HIV positive with red color and HIV negative with bue clolor
}
\description{
Build a cumulative relationships network from the begining of the simulation untill a given time point.
}
\examples{
c.net <- cum.network(datalist = datalist, endtime = 40)
plot.igraph(c.net,layout=layout.fruchterman.reingold)
print(c.net, full = igraph_opt("print.full"))
vertex_attr(c.net, index = V(c.net)) # to print the attributes
}
