% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.df.frac.R
\name{degree.df.frac}
\alias{degree.df.frac}
\title{Fraction of women within specified age group with more than one
sexual partner in the past year.}
\usage{
degree.df.frac(degreedata.df)
}
\arguments{
\item{degreedata.df}{The dataframe that is produced by \code{\link{degree.df.maker}}}
}
\value{
A value, that is, the fraction of women with >1 sexual partner in the last year.
}
\description{
\code{degree.df.frac} Computes the fraction of women within specified age
group with more than one partner from the degreedata dataframe.
}
