% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.df.maker.R
\name{degree.df.maker}
\alias{degree.df.maker}
\title{Degree distribution of recent sexual relationships.}
\usage{
degree.df.maker(dataframe.df, agegroup = c(15, 30), hivstatus = 0,
  survey.time = 40, window.width = 1, gender.degree = "female",
  only.new = TRUE)
}
\arguments{
\item{dataframe.df}{The dataframe that is produced by \code{\link{agemix.df.maker}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that
should be retained, e.g. c(15, 30)}

\item{hivstatus}{HIV status at the time of the survey. Options are 2, means all; 0
means only HIV-negative, 1 means only HIV-positive.}

\item{survey.time}{Time point of the cross-sectional survey.}

\item{window.width}{Time period before the survey e.g 1 year before the survey.}

\item{gender.degree}{The gender of the results}

\item{only.new}{Logical indicator. If TRUE, only relationships that were newly started
during window.width are counted
(i.e. the individual was NEVER in a relationship with these partners before
 the start of the window).
If FALSE, all relationships that were ongoing at some point during the
window.width are counted.}
}
\value{
A dataframe, degree.df that has the columns "ID", "degree", and a fraction of women
with >1 sexual partner in the last year.
}
\description{
Creates a table of two columns, "ID" and "degree" from the df dataframe,
with specific degrees of new relationships or ongoing relationships among women within
specified age group. The degree distribution shows the total number of relationships that
were ongoing (if only.new == FALSE) at some time in the survey window
or that were newly formed (if only.new == TRUE) during the survey window.
}
