% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.distr.fitter.R
\name{degree.distr.fitter}
\alias{degree.distr.fitter}
\title{Fitting parametric distributions to the degree data}
\usage{
degree.distr.fitter(dataframe = degree.df)
}
\arguments{
\item{degree.df}{The dataframe that is produced by \code{\link{degree.df.maker()}}}
}
\value{
A table of the empirical estimate of the three considered goodness-of-fit statistics.
}
\description{
The user fits various parametric distributions to the degree data and
compare their goodness of fit.To choose on the best fitting distribution,
a table is returned giving the empirical estimate of the three considered
goodness-of-fit statistics.
}
\examples{
load(dataframe)
bestfit <- degree.distr.fitter(dataframe = degree.df)

}

