% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.hist.maker.R
\name{degree.hist.maker}
\alias{degree.hist.maker}
\title{Plot the histogram of a degree distribution.}
\usage{
degree.hist.maker(dataframe = degreedata.df)
}
\arguments{
\item{dataframe}{The dataframe that is produced by \code{\link{degree.df.maker}}}
}
\value{
a ggplot2 object
}
\description{
Plot the histogram of the degree distribution that is stored in
 degreedata.df$Degree, after running \code{\link{degree.df.maker}}
}
