% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.wrapper.R
\name{dummy.wrapper}
\alias{dummy.wrapper}
\title{Fast and simple wrapper function to test MICE-assisted calibration functions.}
\usage{
dummy.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of 4 strictly positive input parameter
values, e.g. c(1000, 2, 3, 4)}
}
\value{
A vector of model features
}
\description{
A made-up model with no meaning whatsoever.
}
\examples{
input.vector <- c(1000, 2, 3, 4)
dummy.features <- dummy.wrapper(input.vector)
}
