% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAAA.F.c.wrapper.R
\name{EAAA.F.c.wrapper}
\alias{EAAA.F.c.wrapper}
\title{Wrapper function for running simpact simulations for the MaxART EAAA simulation study}
\usage{
EAAA.F.c.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of random seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
F.c. is the Future counterfactual scenario: until 2031.75 under conservative ART programme
}
