% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAAA.P.a.wrapper.R
\name{EAAA.P.a.wrapper}
\alias{EAAA.P.a.wrapper}
\title{Wrapper function for running simpact simulations for the MaxART EAAA simulation study}
\usage{
EAAA.P.a.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of random seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
P.a. is the Past alternative scenario: until 2016.75 under observed ART programme
}
