% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAAA.revised.F.a.wrapper.R
\name{EAAA.revised.F.a.wrapper}
\alias{EAAA.revised.F.a.wrapper}
\title{Wrapper function for running simpact simulations for the MaxART EAAA simulation study}
\usage{
EAAA.revised.F.a.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of random seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
F.a. is the Future alternative scenario: until 2031.75 under observed ART programme. This file is for the revised analysis: exponential time till ART non-retention, no msm, extra params for increased rate of HIV diagnosis after 2012, and quarterly logging
}
