% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAAA.revised.wrapper.R
\name{EAAA.revised.wrapper}
\alias{EAAA.revised.wrapper}
\title{Wrapper function for running simpact simulations for the MaxART EAAA simulation study}
\usage{
EAAA.revised.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of random seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
This wrapper function includes updates compared to EAAA.wrapper. The initial EAAA.wrapper was used for the report as submitted on 20 August 2018. This revised wrapper has additional parameters to allow for a better fit to the ART coverage data, and does not allow for MSM relationships. It also has an exponential distribution of times till ART non-retention, compared to a fixed distribution with min=1000 and max=2000, which essentially meant that there was no drop-out.
}
