% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAAA.SciRep.revision2.calibration.wrapper.R
\name{EAAA.SciRep.revision2.calibration.wrapper}
\alias{EAAA.SciRep.revision2.calibration.wrapper}
\title{Wrapper function for running simpact simulations for the calibration step in the MaxART EAAA simulation study}
\usage{
EAAA.SciRep.revision2.calibration.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of random seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
Wrapper function for running simpact simulations for the calibration step in the MaxART EAAA simulation study
}
