% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAAA.wrapper.R
\name{EAAA.wrapper}
\alias{EAAA.wrapper}
\title{Wrapper function for running simpact simulations for the MaxART EAAA simulation study}
\usage{
EAAA.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of random seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
A short description here...
}
