% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi2tree2.R
\name{epi2tree2}
\alias{epi2tree2}
\title{Build a transmission tree object from a transmission network data frame}
\usage{
epi2tree2(epi)
}
\arguments{
\item{epi}{The phylo object produced by \code{\link{transmission.network.builder}}.}
}
\value{
Transmission tree object (a phylo object)
}
\description{
Build a transmission tree object from a transmission network data frame
}
\examples{

transm.ls <- transmission.network.builder(datalist = datalist, endpoint = 40)
epi <- transm.ls[[1]]
tree0 <- epi2tree2(transnetwork = epi) # transmission tree of seeds 1

}
