% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.function.R
\name{err.function}
\alias{err.function}
\title{We expect this to grow and the returned summary statistics to include a lot more.
Will include an Ignore variable to each of the summary statistics. Only computed if IGNORE is set to FALSE}
\usage{
err.function(e)
}
\arguments{
\item{the}{output message from simpact.run \code{\link{readthedata()}}}
}
\value{
a string indicating that the program terminated premature
}
\description{
After the simpact simulation a number of statistics can be computed to summarise the output
}
\details{
art.intro list should include the time when the diagnosis to take effect.

The function take in a list of paramenters set at intervals of cd4 count threshhold used to
decide if a person can be offerd treatment
}
\examples{
out.test <- tryCatch(simpact.run(), error = errFunction)
}
