% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv.infected.group.R
\name{hiv.infected.group}
\alias{hiv.infected.group}
\title{Calculate the number of HIV infections that occured between two time points.}
\usage{
hiv.infected.group(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(20, 30), site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{HIV infection within this agegroup.}

\item{timewindow}{HIV transmission that took place between this simulation time e.g timewindow = c(20, 30)}

\item{site}{Select only the particular site from the study, if all ignore site/use all sites.}
}
\value{
data.frame with number of infection aggregated by Gender
}
\description{
Calculate the number of HIV infections that occured between two time points.
}
\examples{
data(datalist)
hiv.infected <- hiv.infected.group(datalist = datalist,
agegroup=c(15, 30), timewindow=c(20, 40), site="All")
hiv.infected

}
