% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIVinNetworks.calibration.wrapper.R
\name{HIVinNetworks.calibration.wrapper}
\alias{HIVinNetworks.calibration.wrapper}
\title{Wrapper function for running simpact simulations for the calibration step in the HIV in Networks project}
\usage{
HIVinNetworks.calibration.wrapper(index, list_param)
}
\arguments{
\item{index}{index}

\item{list_param}{list of vectors of random number seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
Wrapper function for running simpact simulations for the calibration step in the HIV in Networks project
}
