% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPC_MaC.R
\name{HPC_MaC}
\alias{HPC_MaC}
\title{HPC-version of MICE-assisted calibration}
\usage{
HPC_MaC(targets.empirical = dummy.targets.empirical,
  previous.experiments = input.output.df, file.name, RMSD.tol.max = 2,
  min.givetomice = 12, n.experiments = 48,
  strict.positive.params = c(1, 3), probability.params = 0,
  method = "norm", predictorMatrix = "complete", maxit = 50)
}
\arguments{
\item{targets.empirical}{The vector of target features}

\item{previous.experiments}{Dataframe that stores completed experiments
and their associated model features}

\item{RMSD.tol.max}{Tolerance for the root mean squared distance between
target features and model output features}

\item{min.givetomice}{Minimal number of observations in the training dataset
to which MICE is applied}

\item{n.experiments}{Number of proposed experiments to be produced}

\item{strict.positive.params}{Vector of indices that indicate which of the
input parameters are strictly positive. Set to zero if there are no such
parameters.}

\item{probability.params}{Vector of indices that indicate which of the input
parameters are strictly between 0 and 1. Set to zero if there are no such
parameters.}

\item{method}{Method used by MICE. E.g. "norm" or "rf"}

\item{predictorMatrix}{Can be "complete", "LASSO", or a user-defined matrix
of indices that indicate which variables are included in the chained
equations in MICE}

\item{maxit}{The maxit argument used in MICE (number of times that the
chained equations are cycled through)}
}
\value{
A dataframe with proposed input parameter values to match a vector of
  target features.
}
\description{
Produces a csv file of proposed experiments (input parameter values) to match
a vector of target features.
}
\details{
This version starts from a dataframe of completed experiments and their
associated model features.
}
