% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDs.indiv.seq.gender.age.group.fun.R
\name{IDs.indiv.seq.gender.age.group.fun}
\alias{IDs.indiv.seq.gender.age.group.fun}
\title{Subset sequences based on gender ratio and age groups of individuals}
\usage{
IDs.indiv.seq.gender.age.group.fun(simpact.trans.net = simpact.trans.net,
  limitTransmEvents = 3, perc.men = 30, age.men = c(15, 60),
  age.women = c(15, 40))
}
\arguments{
\item{simpact.trans.net}{Transmission networks computed by \code{\link{transmission.network.builder()}}}

\item{limitTransmEvents}{Choose transmission network with at least this  amount of individuals}

\item{perc.men}{Precentage of men to consider when selecting the sequences}

\item{age.men}{Age group for men}

\item{age.women}{Age group for women}
}
\value{
A list with a vector of ID's of selected sequences, number of men, and women for the selcted sequences and the ratio between men and women among the selcted sequences because in some settings we may have less or more female or male individuals
}
\description{
Subset sequences based on gender ratio and age groups of individuals
}
