% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDs.indiv.seq.gender.fun.R
\name{IDs.indiv.seq.gender.fun}
\alias{IDs.indiv.seq.gender.fun}
\title{Subset sequences based on gender ratio and age of individuals}
\usage{
IDs.indiv.seq.gender.fun(simpact.trans.net = simpact.trans.net,
  limitTransmEvents = 7, perc.men = 50, seq.cov = 35, age.limit = 65)
}
\arguments{
\item{simpact.trans.net}{Transmission networks computed by \code{\link{transmission.network.builder()}}}

\item{limitTransmEvents}{Choose transmission network with at least this  amount of individuals}

\item{perc.men}{Precentage of men to consider when selecting the sequences}

\item{seq.cov}{Sequece coverage}

\item{age.limit}{Age limit for all individuals}
}
\value{
A list with a vector of ID's of selected sequences, number of men, and women for the selcted sequences and the ratio between men and women among the selcted sequences because in some settings we may have less or more female or male individuals
}
\description{
Subset sequences based on gender ratio and age of individuals
}
