% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDs.Seq.Age.Groups.R
\name{IDs.Seq.Age.Groups}
\alias{IDs.Seq.Age.Groups}
\title{Random subset of sequences based on gender ratio and age groups of individuals}
\usage{
IDs.Seq.Age.Groups(simpact.trans.net = simpact.trans.net,
  limitTransmEvents = 7, timewindow = c(10, 40), seq.cov = 70,
  seq.gender.ratio = 0.7, age.group.15.25 = c(15, 25),
  age.group.25.40 = c(25, 40), age.group.40.50 = c(40, 50))
}
\arguments{
\item{simpact.trans.net}{Transmission networks computed by \code{\link{transmission.network.builder()}}}

\item{limitTransmEvents}{Choose transmission network with at least this  amount of individuals}

\item{timewindow}{Time interval in which sampling are considered}

\item{seq.cov}{Sequece coverage}

\item{seq.gender.ratio}{Gender imbalance ratio (women/(women+men))}

\item{age.group.15.25}{Age group}

\item{age.group.25.40}{Age group}

\item{age.group.40.50}{Age group}
}
\value{
A vector of selected IDs
}
\description{
Random subset of sequences based on gender ratio and age groups of individuals
}
