% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDs.Seq.Random.R
\name{IDs.Seq.Random}
\alias{IDs.Seq.Random}
\title{Completely random subset of sequences of individuals}
\usage{
IDs.Seq.Random(simpact.trans.net = simpact.trans.net,
  limitTransmEvents = 7, timewindow = c(10, 40), seq.cov = 70,
  age.limit = 65)
}
\arguments{
\item{simpact.trans.net}{Transmission networks computed by \code{\link{transmission.network.builder()}}}

\item{limitTransmEvents}{Choose transmission network with at least this  amount of individuals}

\item{timewindow}{Time interval in which sampling are considered}

\item{seq.cov}{Sequece coverage}

\item{age.limit}{Age limit for all individuals}
}
\value{
A vector of selected IDs
}
\description{
Completely random subset of sequences of individuals
}
