% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.calculator.R
\name{incidence.calculator}
\alias{incidence.calculator}
\title{Calculate HIV incidence, overall and stratified.}
\usage{
incidence.calculator(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(20, 30))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that should be retained, e.g. c(15, 30)}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound) that should be retained, e.g. c(20, 30)}
}
\value{
a dataframe with cases, exposure time, incidence estimate and surrounding confidence bounds,
for the specified time window and age group, overall, and stratified by gender
}
\description{
Calculate the HIV incidence in a time window and for specific age groups and gender strata.
}
\examples{
incidence.df <- incidence.calculator(datalist = datalist, agegroup = c(15, 30), timewindow = c(20, 30))

}

