% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.vector.creator.R
\name{incidence.vector.creator}
\alias{incidence.vector.creator}
\title{Create a vector of HIV incidence measurements.}
\usage{
incidence.vector.creator(datalist = datalist, agegroup = c(15, 50))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper
bound) that should be retained, e.g. c(15, 30)}
}
\value{
A dataframe with HIV incidence estimates, for the specific age group.
}
\description{
Calculate the HIV incidence, for specific age groups, in yearly intervals,
from one until the end of the simulation.
}
\examples{
data(datalist)
inc.vector.df <- incidence.vector.creator(datalist = datalist, agegroup = c(15,50))
inc.vector.df

}
