% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.output.R
\docType{data}
\name{input.output.df}
\alias{input.output.df}
\title{Input and output dataset for MICE.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 10 columns.}
\usage{
data(input.output.df)

  ormat A dataframe with 48 rows and 10 variables:

  \describe{ \item{seed}{A unique id for each experiment} \item{x.i}{input
  parameter values} \item{y.i}{model features} }
}
\description{
A  dataframe of input parameters and model features from a wave of
experiments.
}
\keyword{datasets}
