% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.params.creator.R
\name{input.params.creator}
\alias{input.params.creator}
\title{Create a default parameter configuration list.}
\usage{
input.params.creator(population.eyecap.fraction = 0.5,
  population.simtime = 40, population.numwomen = 200,
  population.nummen = 200, periodiclogging.interval = 1,
  periodiclogging.starttime = 0, syncrefyear.interval = 1,
  syncpopstats.interval = 1, hivseed.time = 10, hivseed.type = "amount",
  hivseed.amount = 10, hivseed.age.min = 20, hivseed.age.max = 30,
  conception.alpha_base = -3, conception.alpha_agewoman = 0,
  dissolution.alpha_0 = 0.1, dissolution.alpha_4 = -0.05,
  formation.hazard.type = "agegapry",
  formation.hazard.agegapry.gap_factor_man_const = 0,
  formation.hazard.agegapry.gap_factor_woman_const = 0,
  person.agegap.man.dist.type = "normal",
  person.agegap.woman.dist.type = "normal",
  person.agegap.man.dist.normal.mu = -4,
  person.agegap.woman.dist.normal.mu = -4,
  person.agegap.man.dist.normal.sigma = 3,
  person.agegap.woman.dist.normal.sigma = 3,
  formation.hazard.agegapry.baseline = 0,
  formation.hazard.agegapry.numrel_scale_man = 0,
  formation.hazard.agegapry.numrel_scale_woman = 0,
  formation.hazard.agegapry.gap_agescale_man = 0.3,
  formation.hazard.agegapry.gap_agescale_woman = 0.3,
  person.eagerness.dist.type = "gamma",
  formation.hazard.agegapry.eagerness_sum = 1,
  transmission.param.a = -1.0352239, transmission.param.b = -89.339994,
  transmission.param.c = 0.4948478, transmission.param.d1 = 0,
  transmission.param.d2 = 0, transmission.param.f1 = log(5),
  transmission.param.f2 = log(log(2.5)/log(5))/5, person.vsp.toacute.x = 10,
  person.vsp.toaids.x = 7, person.vsp.tofinalaids.x = 12,
  formation.hazard.agegapry.numrel_man = -0.2,
  formation.hazard.agegapry.numrel_woman = -0.2,
  formation.hazard.agegapry.numrel_diff = -0.1,
  person.eagerness.dist.gamma.a = 0.125, person.eagerness.dist.gamma.b = 8,
  formation.hazard.agegapry.eagerness_diff = -1,
  formation.hazard.agegapry.gap_factor_man_exp = -0.2,
  formation.hazard.agegapry.gap_factor_woman_exp = -0.2,
  formation.hazard.agegapry.gap_factor_man_age = 0.05,
  formation.hazard.agegapry.gap_factor_woman_age = 0.05,
  formation.hazard.agegapry.meanage = -0.1)
}
\arguments{
\item{conception.alpha_base}{Baseline parameter for conception rate. (-3)}

\item{conception.alpha_agewoman}{The Effect of the age of the woman on the conception rate. (0)}

\item{dissolution.alpha_0}{Baseline parameter for relationship dissolution rate. (0.1)}

\item{dissolution.alpha_4}{Effect of increasing mean age of the  couple on  relationship dissolution rate (-0.05)}

\item{formation.hazard.type}{Type of hazard function for relationship formation. Choose between "simple", "agegap" and "agegapry".}

\item{formation.hazard.agegapry.gap_factor_man_const}{Baseline parameter for men (0).}

\item{formation.hazard.agegapry.gap_factor_woman_const}{Baseline parameter for women (0).}

\item{person.agegap.man.dist.type}{Distribution of preferred age differences for men ("normal")}

\item{person.agegap.woman.dist.type}{Distribution of preferred age differences for women ("normal")}

\item{person.agegap.man.dist.normal.mu}{Mean of preferred age differences distribution for men (-4)}

\item{person.agegap.woman.dist.normal.mu}{Mean of preferred age differences distribution for women (-4)}

\item{person.agegap.man.dist.normal.sigma}{Standard deviation of preferred age differences distribution for men (3)}

\item{person.agegap.woman.dist.normal.sigma}{Standard deviation of preferred age differences distribution for women (3)}

\item{formation.hazard.agegapry.baseline}{Baseline parameter for relationship formation rate (0)}

\item{formation.hazard.agegapry.numrel_scale_man}{Effect of the number of relationships on the penalty for deviation from preferred age difference for men (0)}

\item{formation.hazard.agegapry.numrel_scale_woman}{Effect of the number of relationships on the penalty for deviation from preferred age difference for women (0)}

\item{formation.hazard.agegapry.gap_agescale_man}{Effect of male age on preferred age difference (~ 1 - slope in regression model FemaleAge ~ MaleAge) (0.3)}

\item{formation.hazard.agegapry.gap_agescale_woman}{Effect of male age on preferred age difference (~ 1 - slope in regression model FemaleAge ~ MaleAge) (0.3)}

\item{person.eagerness.dist.type}{Type of distribution for "eagerness" (~ sexual activity) in the population ("gamma")}

\item{formation.hazard.agegapry.eagerness_sum}{Effect of the sum of eagerness values in both partners on the relationship formation rate (1)}

\item{transmission.param.a}{Baseline parameter for HIV transmission rate in serodiscordant couples (-1.0352239)}

\item{transmission.param.b}{Parameter "b" for the linear component of the effect of viral load on the HIV transmission rate in serodiscordant couples (-89.339994)}

\item{transmission.param.c}{Parameter "c" for the exponential component of the effect of viral load on the HIV transmission rate in serodiscordant couples (0.4948478)}

\item{transmission.param.d1}{Parameter "d1" for effect of number of ongoing relationships the the male partner (coital dilution) (0)}

\item{transmission.param.d2}{Parameter "d2" for effect of number of ongoing relationships the the female partner (coital dilution) (0)}

\item{transmission.param.f1}{Effect of youngest age on HIV susceptibility (log(5) ~1.6 such that the hazard is x 5 in 15 year olds)}

\item{transmission.param.f2}{Effect of female age on HIV susceptibility (log(log(2.5) / log(5)) / 5 ~-0.11 such that the hazard is x 2.5 in 20 year olds, compared to the reference (>>25 year olds)}

\item{person.vsp.toacute.x}{Effect of acute versus chronic HIV infection on infectiousness (10)}

\item{person.vsp.toaids.x}{Effect of "initial" AIDS stage versus chronic HIV infection on infectiousness (7)}

\item{person.vsp.tofinalaids.x}{Effect of "final" AIDS stage versus chronic HIV infection on infectiousness (12)}

\item{formation.hazard.agegapry.numrel_man}{Effect of number of ongoing relationships on the relationship formation rate for men (-0.2)}

\item{formation.hazard.agegapry.numrel_woman}{Effect of number of ongoing relationships on the relationship formation rate for women (-0.2)}

\item{formation.hazard.agegapry.numrel_diff}{Effect of absolute difference in number of ongoing relationships on the relationship formation rate (-0.1)}

\item{person.eagerness.dist.gamma.a}{Shape parameter (kappa parameter on wiki page) of Gamma distribution of "eagerness" (0.125)}

\item{person.eagerness.dist.gamma.b}{Scale parameter (theta parameter on wiki page) of Gamma distribution of "eagerness" (8)}

\item{formation.hazard.agegapry.eagerness_diff}{Effect of absolute difference in eagerness on the relationship formation rate (-1)}

\item{formation.hazard.agegapry.gap_factor_man_exp}{Effect of deviation from preferred age difference on the relationship formation rate for men (-0.2)}

\item{formation.hazard.agegapry.gap_factor_woman_exp}{Effect of deviation from preferred age difference on the relationship formation rate for women (-0.2)}

\item{formation.hazard.agegapry.gap_factor_man_age}{Funnel effect in men (0.05)}

\item{formation.hazard.agegapry.gap_factor_woman_age}{Funnel effect in women (0.05)}

\item{formation.hazard.agegapry.meanage}{Effect of increasing mean age of the candidate couple on the relationship formation rate (-0.1)}
}
\value{
a list of model parameters that can be used as input for simpact.run()
}
\description{
Create a default parameter configuration list as a starting point for a parameter optimisation algorithm.
At this stage, only the most relevant parameters are modifiable.
}
\examples{
cfg.list <- input.params.creator(conception.alpha_base = -3,
formation.hazard.type = "agegapry",
formation.hazard.agegapry.numrel_man = -1,
formation.hazard.agegapry.numrel_woman = -1, ...)
}

