% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.introduced.R
\name{intervention.introduced}
\alias{intervention.introduced}
\title{We need to decide if we need to include all the intervention parameters here?
log viralload, etc}
\usage{
intervention.introduced(simulation.type = "simpact-cyan")
}
\arguments{
\item{simulation.type}{An indication of which of simpact simulation to perform.}
}
\value{
a list of lists containing intervention at different times.
The first list sets the diagnosis.baseline value.
}
\description{
Introduce the intervention years and monitoring time for the cd4. Note that
art.intro list should include the time when the diagnosis to take effect.
}
\details{
The function take in a list of paramenters set at intervals of cd4 count threshhold used to
decide if a person can be offerd treatment
}
\examples{
intervention.introduced <- intervention.introduced(simulation.type = "simpact-cyan")
}
