% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.years.lived.R
\name{life.years.lived}
\alias{life.years.lived}
\title{A function that returns the total number of life-years lived between two points in simulation
time for a particular age group and gender}
\usage{
life.years.lived(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(15, 30), site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{alive people within this agegroup.}

\item{timewindow}{alive people within this simulation time e.g timewindow = c(15, 30).}

\item{site}{Select only the particular site from the study, if all ignore site/use all sites.}
}
\value{
the total number of life-years lived aggregated by gender.
}
\description{
A function that returns the total number of life-years lived between two points in simulation
time for a particular age group and gender
}
\examples{
data(datalist)
life.years.lived <- life.years.lived(datalist = datalist, agegroup=c(15, 30),
timewindow=c(15, 30), site="All")
life.years.lived

}
