% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaC.explorer.R
\name{MaC.explorer}
\alias{MaC.explorer}
\title{Distribute simpact runs from MaC output over multiple cores}
\usage{
MaC.explorer(model = model, MaC.output = test.MaC, n_cluster = n_cluster)
}
\arguments{
\item{model}{Wrapper function for running simpact (simpact.wrapper)}

\item{MaC.output}{Output from \link{MaC} function.}

\item{n_cluster}{Number of cores available for parallel running of Simpact}
}
\value{
a matrix of model features and the seed of the random number
  generator
}
\description{
A short description here...
}
