% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaC.weighted.R
\name{MaC.weighted}
\alias{MaC.weighted}
\title{MICE-assisted calibration with weights}
\usage{
MaC.weighted(targets.empirical = dummy.targets.empirical,
  RMSD.tol.max = 2, min.givetomice = 64, n.experiments = 256, lls,
  uls, model = VEME.wrapper, strict.positive.params, probability.params,
  inside_prior = TRUE, method = "norm", predictorMatrix = "complete",
  maxit = 50, maxwaves = 4, n_cluster = n_cluster)
}
\arguments{
\item{targets.empirical}{The vector of target features}

\item{RMSD.tol.max}{Tolerance for the root mean squared distance between
target features and model output features}

\item{min.givetomice}{Minimal number of observations in the training dataset
to which MICE is applied}

\item{n.experiments}{Number of parameter combinations in each wave of model
runs}

\item{lls}{Vector of lower limits of the prior distribution of input
parameter values}

\item{uls}{Vector of upper limits of the prior distribution of input
parameter values}

\item{model}{Wrapper function for the simulation model. See details for a
description of the required format.}

\item{strict.positive.params}{Vector of indices that indicate which of the
input parameters are strictly positive. Set to zero if there are no such
parameters.}

\item{probability.params}{Vector of indices that indicate which of the input
parameters are strictly between 0 and 1. Set to zero if there are no such
parameters.}

\item{inside_prior}{TRUE by default. If FALSE, parameter sampling is not
restricted to the initial ranges of the prior distribution during the
subsequent algorithm steps.}

\item{method}{Method used by MICE. E.g. "norm" or "rf"}

\item{predictorMatrix}{Can be "complete", "LASSO", or a user-defined matrix
of indices that indicate which variables are included in the chained
equations in MICE}

\item{maxit}{The maxit argument used in MICE (number of times that the
chained equations are cycled through)}

\item{maxwaves}{The maximum number of waves of model runs}

\item{n_cluster}{The number of cores available for parallel model runs}
}
\value{
A list with multiple waves of proposed input parameter values to
  match a vector of target features.
}
\description{
Produces a list with multiple waves of proposed input parameter values to
match a vector of target features.
}
\details{
The \code{model} wrapper function for the simulaton model must have a vector
of model input parameter values as its one and only  argument. Furthermore,
it must return a vector of model features. These model features are then
compared against the target features.
}
