% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.age.pos.vector.creator.R
\name{mean.age.pos.vector.creator}
\alias{mean.age.pos.vector.creator}
\title{Create a vector of mean ages among the HIV positive people}
\usage{
\method{mean}{age.pos.vector.creator}(datalist = datalist,
  timevector = datalist$itable$hivseed.time[1]:datalist$itable$population.simtime[1])
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timevector}{Vector of time point at which the mean age among HIV
positive people must be calculated.}
}
\value{
A vector of mean ages among the HIV positive people.
}
\description{
Calculate the HIV prevalence, for specific age groups, in yearly intervals,
from zero until the end of the simulation.
}
\examples{
data(datalist)
mean.age.pos.vector <- mean.age.pos.vector.creator(datalist = datalist, timevector = 10:60)
prev.vector.df

}
