% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.fit.R
\name{mice.fit}
\alias{mice.fit}
\title{MICE with weights}
\usage{
mice.fit(data, m = 5, method = "norm", predictorMatrix = (1 - diag(1,
  ncol(data))), where = is.na(data), visitSequence = NULL,
  form = vector("character", length = ncol(data)),
  post = vector("character", length = ncol(data)),
  defaultMethod = "norm", maxit = 5, diagnostics = TRUE,
  printFlag = TRUE, seed = NA, imputationMethod = NULL,
  defaultImputationMethod = NULL, data.init = NULL, ...)
}
\arguments{
\item{data}{training dataframe}

\item{m}{number of imputations}

\item{method}{string (vector)}

\item{predictorMatrix}{predictorMatrix}

\item{where}{elements in data that are missing}

\item{visitSequence}{defaults is 1 to ncol(data)}

\item{form}{I do not know}

\item{post}{Same}

\item{defaultMethod}{c("pmm", "logreg", "polyreg", "polr")}

\item{maxit}{How many cycles through chained equations}

\item{diagnostics}{Do you want to see diagnostics}

\item{printFlag}{Do you want stuff printed}

\item{seed}{seed value to make it reproducible}

\item{imputationMethod}{Do not know}

\item{defaultImputationMethod}{Same}

\item{data.init}{Same}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Produces a list with imputations with inverse probability weights.
}
