% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.norm.fit.R
\name{mice.impute.norm.fit}
\alias{mice.impute.norm.fit}
\title{adjusted mice norm fitter}
\usage{
mice.impute.norm.fit(y, ry, x, wy = NULL, ...)
}
\arguments{
\item{y}{y}

\item{ry}{ry}

\item{x}{x}

\item{wy}{wy}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Augmented to also return details of the regression model and residual error sampled.
}
