% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.parallel.R
\name{mice.parallel}
\alias{mice.parallel}
\title{Run mice in parallel}
\usage{
mice.parallel(mice.model = mice.wrapper,
  df.give.to.mice = df.give.to.mice, m = 1,
  predictorMatrix = predictorMatrix, method = "norm",
  defaultMethod = "norm", maxit = 5, printFlag = FALSE,
  seed_count = 0, n_cluster = 8, nb_simul = n.experiments)
}
\arguments{
\item{mice.model}{Wrapper function for mice (mice.wrapper)}

\item{df.give.to.mice}{Training dataset to give to mice}

\item{m}{Number of imputations}

\item{predictorMatrix}{Predictor matrix}

\item{method}{Method for imputations}

\item{defaultMethod}{Default method for imputations}

\item{maxit}{Number of times we cycle through chained equations}

\item{printFlag}{print flag TRUE of FALSE}

\item{seed_count}{Origin of random number seed}

\item{n_cluster}{Number of cores available for parallel running of mice}

\item{nb_simul}{Number of guesses to be produced by mice}
}
\value{
a matrix of mice guesses
}
\description{
A short description here...
}
