% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.wrapper.R
\name{mice.wrapper}
\alias{mice.wrapper}
\title{Wrapper function for mice}
\usage{
mice.wrapper(mice.input.list)
}
\arguments{
\item{mice.input.list}{List with arguments for the mice function}
}
\value{
a matrix of mice guesses
}
\description{
A short description here...
}
