% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persreldf.R
\docType{data}
\name{persreldf}
\alias{persreldf}
\title{Relationship episodes from Simpact Simulation.}
\format{A dataframe with 916306 rows and 23 variables:

  \describe{ \item{relid}{A unique id for each relationship that combines
  the unique ID for the male partner (first) and the female partner (second)
  in the relaitonship} \item{episodeorder}{The order in which episodes
  occure are numbered in ascending order} \item{agerelform}{The age of the
  person when the relationship began. In other words, the age at the
  first episode} \item{pagerelform}{The age of their partner when the
  relationship began.} }}
\source{
\url{http://research.edm.uhasselt.be/jori/simpact/0.19.0/simpact_cyan.html}
}
\usage{
data(persreldf)
}
\description{
A  dataframe from Simpact simulation output that keeps columns from the
relationship table and merges them with person-level attributes. This
dataframe is in a long format where each row/observation represents an
episode from each relationship. Each relationship episode is represented
twice: once for each man in the dataset who was the male partner in that
relationship, and once for each woman who was the female partner in the
relationship. This allows the user to conduct stratified analyses based on
gender. For relationships where there is more than one episode, there will be
multiple rows.
}
\keyword{datasets}
