% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.AR.groups.fun.agemix.R
\name{phylo.AR.groups.fun.agemix}
\alias{phylo.AR.groups.fun.agemix}
\title{A function that returns age mixing patterns quantities in transmission clusters
in scenarios where individuals are missing at random}
\usage{
phylo.AR.groups.fun.agemix(simpact.trans.net = simpact.trans.net,
  work.dir = work.dir, dirfasttree = dirfasttree,
  sub.dir.rename = sub.dir.rename, limitTransmEvents = 7,
  timewindow = c(30, 40), seq.cov = 70, seq.gender.ratio = 0.7,
  age.group.15.25 = c(15, 25), age.group.25.40 = c(25, 40),
  age.group.40.50 = c(40, 50))
}
\arguments{
\item{simpact.trans.net}{a list of transmission networks produced by \code{\link{transm.network.builder}}}

\item{work.dir}{working directory}

\item{dirfasttree}{directory where is the fastTree tool}

\item{sub.dir.rename}{subdurectory required when we have to run more than one simulations}

\item{limitTransmEvents}{Number of minimum transmission events to be considered in each transmission networks}

\item{timewindow}{Time interval}

\item{seq.cov}{Percentage of individulas considered for this transmission pattern scenario}

\item{seq.gender.ratio}{Gender ratio}

\item{age.group.15.25}{age group between 15 and 25 years old}

\item{age.group.25.40}{age group between 25 and 40 years old}

\item{age.group.40.50}{age group between 40 and 50 years old}
}
\value{
a vector of number of men and women in different age group, number of transmissions within all age groups, and mean and SD of age different between infectors and infectees
}
\description{
A function that returns age mixing patterns quantities in transmission clusters
in scenarios where individuals are missing at random
}
\examples{
w <- phylo.AR.groups.fun.agemix(simpact.trans.net = simpact.trans.net,
                               work.dir = work.dir,
                               dirfasttree = dirfasttree,
                               sub.dir.rename = sub.dir.rename,
                               limitTransmEvents = 7,
                               timewindow = c(30,40),
                               seq.cov = 70,
                               seq.gender.ratio = 0.7,
                               age.group.15.25 = c(15,25),
                               age.group.25.40 = c(25,40),
                               age.group.40.50 = c(40,50))

}
