% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.simpact.parallel.R
\name{phylo.simpact.parallel}
\alias{phylo.simpact.parallel}
\title{Parallelize simpact with phylodynamic component}
\usage{
phylo.simpact.parallel(model = phylo.simpact.wrapper,
  actual.input.matrix = inputmatrix, seed_count = 0, n_cluster = 4)
}
\arguments{
\item{model}{Wrapper function computing epidemiological features}

\item{actual.input.matrix}{Inputs parameters for simpact setup}

\item{seed_count}{Initiation of seed for reproducibility}

\item{n_cluster}{Number of cores}
}
\description{
Parallelize simpact with phylodynamic component
}
