% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.tree.fasttree.bis.R
\name{phylogenetic.tree.fasttree.bis}
\alias{phylogenetic.tree.fasttree.bis}
\title{Construct and calibrate a phylogenetic tree from sequence data}
\usage{
phylogenetic.tree.fasttree.bis(dir.tree = dirfasttree,
  sub.dir.rename = sub.dir.rename, fasttree.tool = "FastTree",
  calendar.dates = "samplingtimes.all.csv",
  simseqfile = "C.Epidemic_seed.seq.bis.sim.nwk.fasta", count.start = 1977,
  endsim = 40)
}
\arguments{
\item{dir.tree}{Direcotry where we find the FastTree tool}

\item{sub.dir.rename}{sub directory where we find sequences and sampling times files and where the simulation outputs will be stored}

\item{fasttree.tool}{Name of the compiled file of FastTree}

\item{calendar.dates}{File containing named sampling calendar times for each sequence}

\item{simseqfile}{File containing sequences}

\item{count.start}{Calendar year when the simulation started}

\item{endsim}{Number of years when the simulation was done}
}
\value{
A time-stamped phylogenetic  tree with annotation of internal nodes dates
}
\description{
Construct and calibrate a phylogenetic tree from sequence data
}
