% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.tree.fasttree.par.R
\name{phylogenetic.tree.fasttree.par}
\alias{phylogenetic.tree.fasttree.par}
\title{Construct phylogenetic tree with FastTree and calibrate same tree with treedater from sequence alignment}
\usage{
phylogenetic.tree.fasttree.par(dir.tree = dirfasttree,
  sub.dir.rename = sub.dir.rename, fasttree.tool = "FastTree",
  calendar.dates = "samplingtimes.all.csv",
  simseqfile = "C.Epidemic_seed.seq.bis.sim.nwk.fasta", count.start = 1977,
  endsim = 40, clust = TRUE)
}
\arguments{
\item{dir.tree}{Direcotry where we find the FastTree tool}

\item{sub.dir.rename}{sub directory where we find sequences and sampling times files and where the simulation outputs will be stored}

\item{fasttree.tool}{Name of the compiled file of FastTree}

\item{calendar.dates}{File containing named sampling calendar times for each sequence}

\item{simseqfile}{File containing sequences}

\item{count.start}{Calendar year when the simulation started}

\item{endsim}{Number of years when the simulation was done}

\item{clust}{Logic parameter TRUE when running simulation on cluster and FALSE on personal computer}
}
\value{
A time-stamped phylogenetic  tree with annotation of internal nodes dates
}
\description{
Construct phylogenetic tree with FastTree and calibrate same tree with treedater from sequence alignment
}
