% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.tree.phangorn.par.R
\name{phylogenetic.tree.phangorn.par}
\alias{phylogenetic.tree.phangorn.par}
\title{Construct phylogenetic tree with Phangorn and calibrate same tree with treedater from sequence alignment}
\usage{
phylogenetic.tree.phangorn.par(simseqfile = sequ.dna)
}
\arguments{
\item{simseqfile}{File containing sequences}
}
\value{
A time-stamped phylogenetic  tree with annotation of internal nodes dates
}
\description{
Construct phylogenetic tree with Phangorn and calibrate same tree with treedater from sequence alignment
}
