% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetictree.trend.R
\name{phylogenetictree.trend}
\alias{phylogenetictree.trend}
\title{Trend of a phylogenetic tree}
\usage{
phylogenetictree.trend(tree = tree)
}
\arguments{
\item{tree}{A tree object}
}
\value{
A list with two vector with same length, which is used for regression analysis
between the number of subtrees and their sizes.
}
\description{
Compute two vectors, one for the number of tips for a given tree (or size of the tree)
and another one for the number of trees (subtrees ) with corresponding tips
}
\examples{
tree <- rtree(23)

examp1 <- phylogenetictree.trend(tree = tree)
x = examp1$num.tree
y = examp1$size.tree
reg <- lm(log(y) ~ log(x))
cozf = coef(reg)

we expect the sizes of subtrees to decrease follwoing power-law hypothetically
power.law.fit = function(x) exp(cozf[[1]] + cozf[[2]] * log(x))
alpha = -cozf[[2]]
R.square = summary(reg)$r.squared
print(paste("Alpha =", round(alpha, 3)))
print(paste("R square =", round(R.square, 3)))

plot(y ~ x, log = "xy", xlab = "Number of subtree (log)", ylab = "Subtree size (log)",
     col = 1, main = "Degree Distribution")
curve(power.law.fit, col = "red", add = T, n = length(y))

}
