% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.growth.calculator.R
\name{pop.growth.calculator}
\alias{pop.growth.calculator}
\title{Calculate population growth rate.}
\usage{
pop.growth.calculator(datalist = datalist, timewindow = c(0, 20))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound) that should be retained, e.g. c(20, 30)}
}
\value{
population growth rate estimate, for the specified time window
}
\description{
Calculate the growth rate of the entire population, averaged over a particular time window.
}
\examples{
growth.rate <- pop.growth.calculator(datalist = datalist, timewindow = c(0, 20))
}

