% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.calculator.R
\name{prevalence.calculator}
\alias{prevalence.calculator}
\title{Calculate HIV prevalence, overall and stratified.}
\usage{
prevalence.calculator(datalist = datalist, agegroup = c(15, 30),
  timepoint = 30)
}
\arguments{
\item{datalist}{The list object that is produced by \code{\link{readthedata}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper 
bound) that should be retained. Should be expressed as a vector of two 
integers. e.g. c(15, 30)}

\item{timepoint}{Point in simulation time at which HIV prevalence should be
calculated}
}
\value{
A data frame with prevalence estimate and surrounding confidence 
  bounds, for the specified time point and age group, overall, and stratified
  by gender.
}
\description{
Produces a data frame that contains the overall, and gender-stratified,
HIV prevalence at a specified point in simulation time and for a specific
age group.
}
\examples{
cfg <- list() 
modeloutput <- RSimpactCyan::simpact.run(configParams = cfg,
destDir = "temp") 
dl <- readthedata(modeloutput) 
prev.df <- prevalence.calculator(datalist = dl, agegroup = c(15, 30), timepoint = 30) 
prev.df

}
