% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.plotter.R
\name{prevalence.plotter}
\alias{prevalence.plotter}
\title{Plot the overall HIV prevalence for the duration of the simulation.}
\usage{
prevalence.plotter(datalist = datalist, agegroup = c(15, 50))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that should be retained, e.g. c(15, 50)}
}
\value{
a ggplot2 object
}
\description{
Plot the overall HIV prevalence, with confidence interval, for the duration of the simulation, for a specific age group.
}
\examples{
prevalence.df <- prevalence.calculator(datalist = datalist, agegroup = c(15, 30), timepoint = 30)

}

