% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.vector.creator.R
\name{prevalence.vector.creator}
\alias{prevalence.vector.creator}
\title{Create a vector of HIV prevalence measurements.}
\usage{
prevalence.vector.creator(datalist = datalist, agegroup = c(15, 50))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper
bound) that should be retained, e.g. c(15, 30)}
}
\value{
A dataframe with HIV prevalence estimates and surrounding confidence
  bounds, for the specific age group.
}
\description{
Calculate the HIV prevalence, for specific age groups, in yearly intervals,
from zero until the end of the simulation.
}
\examples{
data(datalist)
prev.vector.df <- prevalence.vector.creator(datalist = datalist, agegroup = c(15,50))
prev.vector.df

}
