% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readthedata.R
\name{readthedata}
\alias{readthedata}
\title{Read the Simpact output files.}
\usage{
readthedata(modeloutput)
}
\arguments{
\item{modeloutput}{The list object that is produced by
\code{\link[RSimpactCyan]{simpact.run}}.}
}
\value{
A list, containing dataframes for the output of the model run: 
  \describe{
    \item{ptable}{People in the simulation}
    \item{rtable}{Relationship information} 
    \item{etable}{Events} 
    \item{ttable}{HIV treatment episodes} 
    \item{itable}{Input parameters} 
    \item{ltable}{Optional book-keeping log} 
    \item{vltable}{HIV viral load information}
    \item{ftable}{Optional Facility log} 
 }
}
\description{
\code{readthedata} imports the .csv files generated by 
\code{\link[RSimpactCyan]{simpact.run}} in the RSimpactCyan package, and
combines them into one list object. This object is in a format that makes
subsequent analysis easier.
}
\examples{
cfg <- list()
modeloutput <- RSimpactCyan::simpact.run(configParams = cfg, destDir = "temp")
datalist <- readthedata(modeloutput)

}
