% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.study.clients.R
\name{recruit.study.clients}
\alias{recruit.study.clients}
\title{A function that recruits the MaxART clients}
\usage{
recruit.study.clients(datalist = datalist)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}
}
\value{
the ptable datalist with clients tagged as study Yes/No
}
\description{
A function that recruits the MaxART clients
}
\examples{
data(datalist)
datalist.testme <- recruit.study.clients(datalist = datalist.test)
datalist

}
