% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.gender.ratio.R
\name{relationship.gender.ratio}
\alias{relationship.gender.ratio}
\title{Relationship gender ratio}
\usage{
relationship.gender.ratio(datalist = datalist, timewindow = c(20, 40))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound)
that should be retained, e.g. c(20, 40)
data(datalist)
rels.ratio <- relationship.gender.ratio(datalist = datalist, timewindow = c(20, 40))
rels.ratio}
}
\description{
Calculate the relationship gender ratio (ratio of men & women in relationship) over a particular time window.
}
