% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.rate.R
\name{relationship.rate}
\alias{relationship.rate}
\title{Relationship rate}
\usage{
relationship.rate(datalist = datalist, timewindow = c(20, 40))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound)
that should be retained, e.g. c(20, 40)

data(datalist)
rels.rate <- relationship.rate(datalist = datalist, timewindow = c(20, 40))
rels.rate}
}
\description{
Calculate the relationship rate of the entire population, averaged over a particular time window.
}
