% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.fit.R
\name{sampler.fit}
\alias{sampler.fit}
\title{adjusted mice sampler}
\usage{
sampler.fit(data, m, where, imp, blocks, method, visitSequence, predictorMatrix,
  formulas, blots, post, fromto, printFlag, ...)
}
\arguments{
\item{data}{dataframe with missing data in it}

\item{m}{number of cycles I think}

\item{where}{points to missing data}

\item{imp}{number of imputations}

\item{blocks}{blocks}

\item{method}{method}

\item{visitSequence}{Default is 1 to ncol(data)}

\item{predictorMatrix}{predictorMatrix}

\item{formulas}{formulas}

\item{blots}{blots}

\item{post}{post}

\item{fromto}{fromto}

\item{printFlag}{Want a flag}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Augmented to also return details of the regression model and residual error sampled.
}
