% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.univ.fit.R
\name{sampler.univ.fit}
\alias{sampler.univ.fit}
\title{adjusted mice univariate sampler}
\usage{
sampler.univ.fit(data, r, where, type, formula, method, yname, k,
  calltype = "type", user, ...)
}
\arguments{
\item{data}{dataframe with missing data in it}

\item{r}{r}

\item{where}{points to missing data}

\item{type}{type}

\item{formula}{formula}

\item{method}{method}

\item{yname}{yname}

\item{k}{k}

\item{calltype}{calltype}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Augmented to also return details of the regression model and residual error sampled.
}
