% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.summary.params.R
\name{select.summary.params}
\alias{select.summary.params}
\title{Select the summary statistics that you want from the simpact.run.}
\usage{
select.summary.params(..., summary.replace = FALSE)
}
\arguments{
\item{...}{standard summary statistics}

\item{summary.replace}{Set to True if you want to use different statistics}
}
\value{
a list of parameters to be included in the summary statistics of the simulation
}
\description{
The default parameters are all the parameters produced by output.summary.maker.
To have less of these you will need to call the function with summary.replace=TRUE
}
\examples{
target.variables <- select.summary.params()[[1]]
}
