% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.simulation.R
\name{sequence.simulation}
\alias{sequence.simulation}
\title{Simulate the sequences for individual in the transmissiion network using phylosim}
\usage{
sequence.simulation(transtree = tree0, seedSeq = hivSeq, alpha = 0.9,
  rate.list = rate, base.freq = freq)
}
\arguments{
\item{transtree}{Tree (or phylo) object from  \code{\link{trans.network2tree()}}
which respresents the transmisssion tree built from transmission networks
\code{\link{transmNetworkBuilder.baseline()}} or \code{\link{transmNetworkBuilder.diff()}}}

\item{seedSeq}{Input sequence of the seed, it might be a string}

\item{alpha}{Shape parameter of Gamma distribution for rates variation}

\item{r}{A vector of bases rate of substitutions}

\item{f}{A list of bases frequences}
}
\value{
Consensus sequences of recipients for each seed
}
\description{
From the \code{\link{transnetwork2tree()}} we have the object of the transmission tree.
We then can simulate the consesus sequences of individuals using GTR+Gamma model
}
