% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.simulation.seqgen.bis.R
\name{sequence.simulation.seqgen.bis}
\alias{sequence.simulation.seqgen.bis}
\title{Simulate sequences data of individuals in the transmissiion network using seq-gen tool}
\usage{
sequence.simulation.seqgen.bis(dir.seq = dir,
  sub.dir.rename = sub.dir.rename, seq.gen.tool = "seq-gen",
  datalist = datalist, seeds.num = 123, endpoint = 40,
  limitTransmEvents = 3, hiv.seq.file = "hiv.seq.C.pol.j.fasta")
}
\arguments{
\item{dir.seq}{Direcotry where is the compiled seq-gen tool}

\item{sub.dir.rename}{sub directory where the simulation outputs will be stored}

\item{seq.gen.tool}{Name of the file for compiled seq-gen}

\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{seeds.num}{Seed number for reproducability}

\item{endpoint}{Only transmission events that took place before this point in simulation time}

\item{limitTransmEvents}{Minimum number of individuals in that transmission network (one transmission network per HIV seeding individual)}

\item{hiv.seq.file}{File containing seed HIV sequences in FASTA format, it might be in same repository as seq-gen}
}
\value{
Files of sequences data, sampling dates, and transmission trees
}
\description{
Simulate sequences data of individuals in the transmissiion network using seq-gen tool
}
