% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.simulation.seqgen.par.R
\name{sequence.simulation.seqgen.par}
\alias{sequence.simulation.seqgen.par}
\title{Simulate sequence data of individuals in the transmissiion network using seq-gen tool}
\usage{
sequence.simulation.seqgen.par(dir.seq = dir,
  sub.dir.rename = sub.dir.rename, seq.gen.tool = "seq-gen",
  simpact.trans.net = simpact.trans.net, seeds.num = 123,
  endpoint = 40, limitTransmEvents = 3,
  hiv.seq.file = "hiv.seq.C.pol.j.fasta", clust = TRUE)
}
\arguments{
\item{dir.seq}{Direcotry where is the compiled seq-gen tool, sequences data may also be stored there}

\item{sub.dir.rename}{sub directory where the simulation outputs will be stored}

\item{seq.gen.tool}{Name of the file for compiled seq-gen}

\item{simpact.trans.net}{Transmission network produced by \code{\link{transmission.network.builder()}}}

\item{seeds.num}{Seed number for reproducability}

\item{endpoint}{Only transmission events that took place before this point in simulation time}

\item{limitTransmEvents}{Minimum number of individuals in that transmission network (one transmission network per HIV seeding individual)}

\item{hiv.seq.file}{File containing seed HIV sequences in FASTA format, it might be in same repository as seq-gen}
}
\value{
Files of sequences data, sampling dates, and transmission trees
}
\description{
Simulate sequence data of individuals in the transmissiion network using seq-gen tool
}
