% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpact.config.inputs.add.sample.R
\name{simpact.config.inputs.add.sample}
\alias{simpact.config.inputs.add.sample}
\title{Resampling of Simpact config input parameters to be estimated.}
\usage{
simpact.config.inputs.add.sample(datalist = datalist,
  resample.points = 10, set.seed.new = 1, ...)
}
\arguments{
\item{datalist}{the dataframe produced by simpact.wrapper.easyABC.run}

\item{resample.points}{Number of points to be added standard varied simpact parameters}

\item{set.seed.new}{In case cluster, you need to be using different seed for resampling}

\item{...}{a set of parameter boundaries to be used for new re-sampling}
}
\value{
dataframe with resampled parameter values to be used with simpact.run
}
\description{
This is the first step in setting up for the multivator/emulator parameter estimation.
The user needs to list all the parameters that needs to be estimated and also specify
the boundaries for these parameters
The dataframe that is returned includes the design matrix ([0,1] random numbers for
each parameter) and the specific parameter value that will be used to run simpact
as used in the initial set up
}
