% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpact.config.inputs.par.select.R
\name{simpact.config.inputs.par.select}
\alias{simpact.config.inputs.par.select}
\title{Simpact config input parameters to be estimated.}
\usage{
simpact.config.inputs.par.select(datalist = datalist)
}
\arguments{
\item{datalist}{the selected parameters that produced minimal sum of squares}
}
\value{
dataframe with varied parameter values to be used with simpact.run
}
\description{
This is the first step in setting up for the multivator/emulator parameter estimation.
The user needs to list all the parameters that needs to be estimated and also specify
the boundaries
for these parameters
The dataframe that is returned includes the design matrix ([0,1] random numbers for
each parameter) and the specific parameter value that will be used to run simpact
}
