% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpact.config.inputs.R
\name{simpact.config.inputs}
\alias{simpact.config.inputs}
\title{Simpact config input parameters to be estimated.}
\usage{
simpact.config.inputs(design.points = 10, resample.count = 1, ...)
}
\arguments{
\item{design.points}{the number simulation rows that needs to be perfomed by the simpact.run}

\item{resample.count}{replace the samples with these number of resample}

\item{...}{standard varied simpact parameters}
}
\value{
dataframe with varied parameter values to be used with simpact.run
}
\description{
This is the first step in setting up for the multivator/emulator parameter estimation.
The user needs to list all the parameters that needs to be estimated and also specify
the boundaries for these parameters
The dataframe that is returned includes the design matrix ([0,1] random numbers for
each parameter) and the specific parameter value that will be used to run simpact
}
\examples{
input.varied.params <- simpact.config.inputs(conception.alpha_base = c(-3.6, -1.2),
formation.hazard.agegapry.baseline = c(1.5, 3), person.eagerness.man.dist.gamma.a = c(0.1, 2),
person.eagerness.man.dist.gamma.b = c(5, 60),
formation.hazard.agegapry.numrel_man = c(-1.5, -0.1),
formation.hazard.agegapry.eagerness_diff = c(-0.1, 0),
formation.hazard.agegapry.gap_factor_man_exp = c(-1.5, -0.4),
person.agegap.man.dist.normal.mu = c(0, 4), person.agegap.woman.dist.normal.mu = c(0, 4),
person.agegap.man.dist.normal.sigma = c(0.5,2),person.agegap.woman.dist.normal.sigma =c(0.5,2))

}
