% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpact.errFunction.R
\name{simpact.errFunction}
\alias{simpact.errFunction}
\title{Handling Simpact errors.}
\usage{
simpact.errFunction(e)
}
\arguments{
\item{e}{The error object returned by the Simpact run.}
}
\value{
An empty list.
}
\description{
\code{simpact.errFunction} returns an empty list if the error message of a
problematic Simpact run contains \code{NA}.
}
\examples{
cfg <- list()
results <- tryCatch(simpact.run(configParams = cfg, destDir = "temp"),
                   error = simpact.errFunction)
}
