% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpact.parallel.R
\name{simpact.parallel}
\alias{simpact.parallel}
\title{Distribute simpact runs over multiple cores}
\usage{
simpact.parallel(model = simpact.wrapper,
  actual.input.matrix = matrix(rep(c(1:15), 16), nrow = 16), seed_count = 0,
  n_cluster = 8)
}
\arguments{
\item{model}{Wrapper function for running simpact (simpact.wrapper)}

\item{actual.input.matrix}{Matrix with parameter combinations to be run}

\item{seed_count}{Origin of random number seed}

\item{n_cluster}{Number of cores available for parallel running of Simpact}
}
\value{
a matrix of model features and the seed of the random number
  generator
}
\description{
A short description here...
}
