% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpact.params.boundaries.R
\name{simpact.params.boundaries}
\alias{simpact.params.boundaries}
\title{These are the preset boundaries for most of the parameters that are likely
to be calibrated in simpact.
Most of these parameters will be standard
in \code{\link{simpact.config.inputs}} stage.
The user can set the simulation specific parameters as needed}
\usage{
simpact.params.boundaries(...)
}
\arguments{
\item{...}{is to return all the parameters used
during the \code{\link{simpact.config.inputs}} setup.}
}
\value{
a list of boundaries of parameters to be calibrated
}
\description{
These are the preset boundaries for most of the parameters that are likely
to be calibrated in simpact.
Most of these parameters will be standard
in \code{\link{simpact.config.inputs}} stage.
The user can set the simulation specific parameters as needed
}
\examples{
varied.params.bound <- simpact.params.boundaries()[[1]]
}
