% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpact.wrapper.R
\name{simpact.wrapper}
\alias{simpact.wrapper}
\title{Wrapper function for running simpact simulations}
\usage{
simpact.wrapper(inputvector = input.vector)
}
\arguments{
\item{inputvector}{Vector of random seed and parameter values}
}
\value{
A vector of model features (summary statistics of simulation output)
}
\description{
A short description here...
}
