% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.mrca.matrix.R
\name{time.mrca.matrix}
\alias{time.mrca.matrix}
\title{Matrix of time to the most recent common ancestor (MRCA)}
\usage{
\method{time}{mrca.matrix}(tree = phylo.tree)
}
\arguments{
\item{tree}{A tree (phylo object) in nexus format (a phylogeny)}
}
\value{
A symmetrix matrix where each entry is time where two sequences have the same most recent ancestor
}
\description{
Coumpte time point to the  most recent common ancestor (MRCA) bewteen individuals
}
