% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.point.network.R
\name{time.point.network}
\alias{time.point.network}
\title{Build current relationships network at a given point of time}
\usage{
\method{time}{point.network}(datalist = datalist, time = 30)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{time}{Only relationships present at that time point.}
}
\value{
A graph object for relationships network data.
Men are represented by square, women by sphere, HIV positive with red color and HIV negative with bue clolor
}
\description{
Build current relationships network at a given point of time
}
\examples{
tp.net <- time.point.network(datalist = datalist, time = 30)
plot.igraph(tp.net,layout=layout.fruchterman.reingold)
print(tp.net, full = igraph_opt("print.full"))
vertex_attr(tp.net, index = V(tp.net)) # to print the attributes


}
