% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.wind.network.R
\name{time.wind.network}
\alias{time.wind.network}
\title{Build a cumulative relationships network in a given time interval}
\usage{
\method{time}{wind.network}(datalist = datalist, duration = c(1, 15))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{duration}{Only relationships events that took place between that time interval.}
}
\value{
A graph object for relationships network data.
}
\description{
Build a cumulative relationships network in a given time interval
}
\examples{
tw.net <- time.wind.network(datalist = datalist, duration = c(1,15))
plot.igraph(tw.net,layout=layout.fruchterman.reingold)
print(tw.net, full = igraph_opt("print.full"))
vertex_attr(tw.net, index = V(tp.net)) # to print the attributes
}
