% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transm.network.builder.R
\name{transm.network.builder}
\alias{transm.network.builder}
\title{Build a list with HIV transmission network data.}
\usage{
transm.network.builder(datalist = datalist, endpoint = 30)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{endtime}{Only transmission events that took place before this point in simulation time, are captured in the output.}
}
\value{
a list with the transmission network data, as required by the epi2tree function.
}
\description{
Calculate the HIV prevalence at a point in time, for specific age groups and gender strata.
}
\examples{
transm.ls <- transm.network.builder(datalist = datalist, endpoint = 30)
}

