% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.gender.ratio.R
\name{transmission.gender.ratio}
\alias{transmission.gender.ratio}
\title{Transmission gender ratio}
\usage{
transmission.gender.ratio(datalist = datalist, timewindow = c(20, 40))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound)
that should be retained, e.g. c(20, 40)
data(datalist)
transm.ratio <- transmission.gender.ratio(datalist = datalist, timewindow = c(20, 40))
transm.ratio}
}
\description{
Calculate the transmission gender ratio (ratio of men & women in relationship) over a particular time window.
}
