% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.rate.calculator.R
\name{transmission.rate.calculator}
\alias{transmission.rate.calculator}
\title{Calculate the transmission rate}
\usage{
transmission.rate.calculator(datalist = datalist, timewindow = c(20, 40))
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound)
that should be retained, e.g. c(20, 40)

data(datalist)
transm.rate <- transmission.rate.calculator(datalist = datalist, timewindow = c(20, 40))
transm.rate}
}
\description{
Calculate the transmission rate of the entire population, averaged over a particular time window.
}
