% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmNetworkBuilder.baseline.R
\name{transmNetworkBuilder.baseline}
\alias{transmNetworkBuilder.baseline}
\title{Build a list with HIV transmission networks data.}
\usage{
transmNetworkBuilder.baseline(datalist = datalist, hivseed.time = 10,
  endpoint = 40, population.simtime = 40)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{endtime}{Only transmission events that took place before this point in simulation time, are captured in the output.}
}
\value{
A list with the transmission network data (each of them considering same sampling/removal date), as required by the epi2tree function.
}
\description{
Build a list with HIV transmission networks data.
}
\note{
transm.ls[[1]] is the transmission network (epi object) of seed 1
}
\examples{
transm.ls <- transmNetworkBuilder.baseline(datalist = datalist,endpoint = 40)
}
