% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmNetworkBuilder.diff.R
\name{transmNetworkBuilder.diff}
\alias{transmNetworkBuilder.diff}
\title{Build a list with HIV transmission networks data.}
\usage{
transmNetworkBuilder.diff(datalist = datalist, endpoint = 40,
  population.simtime = 40)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{population.simtime}{Simulation time}

\item{endtime}{Only transmission events that took place before this point in simulation time, are captured in the output.}
}
\value{
a list with the transmission network data ( ach of them considering different sampling/removal dates), as required by the epi2tree function.
}
\description{
Build a list with HIV transmission networks data.
}
\note{
transm.ls[[1]] is the transmission network (epi object) of seed the first seed
}
\examples{
transm.ls <- transmNetworkBuilder.diff(datalist = datalist,endpoint = 40,population.simtime=40)
}
