% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varied.simpact.params.R
\name{varied.simpact.params}
\alias{varied.simpact.params}
\title{Select the parameters that need to be calibrated to meet the target statistics.}
\usage{
varied.simpact.params(..., params.replace = FALSE)
}
\arguments{
\item{...}{standard varied simpact parameters

the \code{\link{simpact.config.inputs}} setup.}

\item{params.replace}{is to return all the parameters used during}
}
\value{
a list of parameters to be calibrated
}
\description{
The default parameters are all the parameters are used during the set
up of \code{\link{simpact.config.inputs}}.
To exclude these and call for new ones, you will need to call the
function with params.replace=TRUE
}
\examples{
input.varied.params.plus <- varied.simpact.params()[[1]]
}
