% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VL.atARTinit.R
\name{vl.atARTinit}
\alias{vl.atARTinit}
\title{A function that returns the total number of people between two points in simulation
time for a particular age group whose VL is within a threshold}
\usage{
vl.atARTinit(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(15, 30), viralload = c(3, 4), site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{alive people within this agegroup.}

\item{timewindow}{alive people within this simulation time e.g timewindow = 30.}

\item{viralload}{log10 set point viral load threshold at the time of ART initiation.}

\item{site}{Select only the particular site from the study, if all ignore site/use all sites.}
}
\value{
the total number of of people whose VL at ART initiation are within a given threshold
}
\description{
A function that returns the total number of people between two points in simulation
time for a particular age group whose VL is within a threshold
}
\examples{
data(datalist)
vl.atARTinit <- vl.atARTinit(datalist = datalist, agegroup = c(15, 30),
timewindow = c(15, 30), viralload = c(3, 4))

}
