% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VL.suppressed.R
\name{vl.suppressed}
\alias{vl.suppressed}
\title{A function that returns the percentage of clients who are alive and are
virally suppressed 6 or more months
after ART initiation, at a paticular point in the simulation time}
\usage{
vl.suppressed(datalist = datalist, timepoint = 30, vlcutoff = 1000,
  lessmonths = 6, site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timepoint}{alive people within this simulation time e.g timewindow = 30.}

\item{vlcutoff}{viral load below this threshold e.g <1000 copies/mL is defined as suppressed}

\item{lessmonths}{months from which time point viral load suppression is sort e.g 6 or
more months after ART initiation}

\item{site}{Select only the particular site from the study, if all ignore site/use all sites.}
}
\value{
the total number and percentage of people who are virally suppressed 6 or more
months after ART initiation at one timepoint
}
\description{
TO REVISE (Need to get the first time that ART was started then reset when
drop out and look at the VL six month after, Ignore everthing else
after this)
}
\examples{
data(datalist)
vl.suppressed <- vl.suppressed(datalist = datalist, timepoint=40,
vlcutoff=1000, lessmonths = 6, site="All")
vl.suppressed

}
