% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VL.suppression.calculator.R
\name{VL.suppression.calculator}
\alias{VL.suppression.calculator}
\title{Calculate viral suppression fraction.}
\usage{
VL.suppression.calculator(datalist = datalist, agegroup = c(15, 30),
  timepoint = 30, vl.cutoff = 1000, site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Boundaries of the age group (lower.bound <= age <
upper.bound) that should be retained, e.g. agegroup = c(15, 30)}

\item{timepoint}{Point in time at which the viral suppression fraction should
be calculated.}

\item{vl.cutoff}{viral load below this threshold e.g <1000 copies/mL is
defined as suppressed}

\item{site}{Select only the particular site from the study, if all ignore
site/use all sites.}
}
\value{
a dataframe with VL suppression estimates and surrounding confidence
  bounds, for the specified time point and age group, overall, and stratified
  by gender.
}
\description{
Calculate the fraction of HIV infected people of a particular age group who
are virally suppressed at a point in time.
}
\examples{
data(datalist)
VL.suppression.df <- VL.suppression.calculator(datalist = datalist,
agegroup = c(15, 30), timepoint = 30, vl.cutoff = 1000, site="All")

}
