% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave1.creator.R
\name{wave1.creator}
\alias{wave1.creator}
\title{Create a csv file with parameter combinations for the first wave of Simpact simulations}
\usage{

  wave1.creator(file.name = "/user/data/gent/vsc400/vsc40070/agemixing/scripts/wave1.csv",
  n.experiments, limits)
}
\arguments{
\item{file.name}{Where the csv file will be stored, and its name}

\item{n.experiments}{Number of parameter combinations to be created}

\item{limits}{list of prior distributions, one per parameter}
}
\value{
a csv file, written to disk
}
\description{
A short description here...
}
