% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.phylo.simpact.parallel.R
\name{wrapper.phylo.simpact.parallel}
\alias{wrapper.phylo.simpact.parallel}
\title{Wrapper function for simpact and phylodynamics}
\usage{
wrapper.phylo.simpact.parallel(inputvector)
}
\arguments{
\item{inputvector}{Vector of model input parameter values}
}
\value{
A vector of model features
}
\description{
This function simulates an HIV epidemic and molecular evolution of HIV across
transmission network
}
