% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.phylosimpact.par.R
\name{wrapper.phylosimpact.par}
\alias{wrapper.phylosimpact.par}
\title{Wrapper for pre and post simpact simulation}
\usage{
wrapper.phylosimpact.par(inputvector, work.dir)
}
\arguments{
\item{inputvector}{Simpact input parameters to setup the simulation}

\item{work.dir}{working directory}
}
\value{
Features
}
\description{
Wrapper for pre and post simpact simulation
}
