% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd4.vl.atARTinit.R
\name{cd4.vl.atARTinit}
\alias{cd4.vl.atARTinit}
\title{A function that returns the total number of people between two points in simulation
time for a particular age group and gender whose VL and CD4count are as indicated}
\usage{
cd4.vl.atARTinit(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(15, 30), viralload = c(3, 4), cd4count = c(350,
  500), site = "All")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{alive people within this agegroup.}

\item{timewindow}{alive people within this simulation time e.g timewindow = 30.}

\item{viralload}{at the time of ART initiation.}

\item{cd4count}{threshold at the time of ART initiation.}

\item{site}{Select only the particular site from the study, if all ignore site/use all sites.}
}
\value{
the total number of of people whose VL and CD4 count at ART initiation are given
}
\description{
A function that returns the total number of people between two points in simulation
time for a particular age group and gender whose VL and CD4count are as indicated
}
\examples{
data(datalist)
cd4.vl.atARTinit <- cd4.vl.atARTinit(datalist = datalist,
agegroup=c(15,40), timewindow=c(15,40), viralload=c(3,4), cd4count=c(350,500), site="All")

}
