% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.df.maker.R
\name{degree.df.maker}
\alias{degree.df.maker}
\title{Degree distribution of persons in recent sexual relationships.}
\usage{
degree.df.maker(df, agegroup = c(15, 30), hivstatus = 0, survey.time = 40,
  window.width = 1, gender.degree = "female", only.new = TRUE)
}
\arguments{
\item{df}{The dataframe that is produced by \code{\link{agemix.df.maker}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper
bound) that should be retained. Should be expressed as a vector of two
integers. e.g. c(15, 30)}

\item{hivstatus}{HIV status at the time of the survey. Options are: 0 = only
HIV-negative, 1 = only HIV-positive, 2 = all}

\item{survey.time}{Numeric integer representing the time point of the
cross-sectional glance at the data}

\item{window.width}{Numeric integer representing how long before the
cross-sectional glance should you look at the data. E.g 1 year before the
survey}

\item{gender.degree}{The gender or the persons in the data. c("male",
"female")}

\item{only.new}{Logical indicator. If TRUE, only relationships that were
newly started during window.width are counted (i.e. the individual was
NEVER in a relationship with these partners before the start of the
window). If FALSE, all relationships that were ongoing at some point during
the window.width are counted.}
}
\value{
A dataframe that has the columns "ID" and "Degree"
}
\description{
Creates a dataframe for the degree distribution of persons specified from the
dataset produced by \code{\link{agemix.df.maker}}. This function allows users
to specify whether they want new relationships or ongoing relationships among
a specified age group, gender, or HIV status of persons in the dataset.
}
\examples{
cfg <- list()
modeloutput <- RSimpactCyan::simpact.run(configParams = cfg, destDir = "temp")
dl <- readthedata(modeloutput)
rels.df <- agemix.df.maker(datalist = dl)
degree.df <- degree.df.maker(rels.df, agegroup = c(20, 30), hivstatus = 1)
}
