% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.modeller.R
\name{pattern.modeller}
\alias{pattern.modeller}
\title{Age-mixing pattern}
\usage{
pattern.modeller(df, agegroup, timepoint, timewindow, start = FALSE)
}
\arguments{
\item{df}{The dataframe that is produced by \code{\link{agedif.df.maker()}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that should be retained, e.g. c(15, 30)}

\item{timepoint}{Point in time at which the age-mixing metrics should be calculated.}

\item{timewindow}{The length of time before the timepoint for which relationships should be included,
e.g. 1, representing one year before the timepoint. This should be a whole number.}

\item{start}{This is a logical indicating that only relationships starting after the beginning of the window
should be used. If start = FALSE relationships could start before the time window.}
}
\value{
a list that contains two elements. One, a subsetted dataframe of relationships that
meet criteria of the function's arguments, including partner age predictions from
the model, and two, a dataframe that contains key age-mixing pattern
outputs from the model.
}
\description{
Models the age-mixing pattern in the population for relationships that
occurred within a specified time window at a specified point in the simulation.
}
\examples{
amp <- pattern.modeller(df = dataframe, agegroup = c(15, 30), timewindow = 1, timepoint = 30)
}

