% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.plotter.R
\name{pattern.plotter}
\alias{pattern.plotter}
\title{Age-mixing pattern plot}
\usage{
pattern.plotter(df)
}
\arguments{
\item{df}{The dataframe that is produced by \code{\link{pattern.modeller()}}}
}
\value{
a large gg plot of the age-mixing pattern
}
\description{
Visualizes the age-mixing pattern in the population.
}
\examples{
amp <- pattern.plotter(df = dataframe)
}

