% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.tree.fasttree.R
\name{phylogenetic.tree.fasttree}
\alias{phylogenetic.tree.fasttree}
\title{COnstruct and calibrate a phylogenetic tree from sequence data}
\usage{
phylogenetic.tree.fasttree(dir.seq = dirseqgen, dir.tree = dirfasttree,
  calendar.dates = "dates.csv", simseqfile = "C.Epidemic.sequence.fasta",
  endsim = 40)
}
\arguments{
\item{dir.seq}{Direcotry where we find sequences and sampling times files}

\item{dir.tree}{Direcotry where we find the FastTree tool}

\item{calendar.dates}{File containing named sampling times for each sequence}

\item{simseqfile}{File containing sequences}

\item{endsim}{End of simulation # place holder for sequence from simpact when assigning real calendar time}
}
\value{
A time stamped phylogenetic  tree with annotation of internal nodes dates
}
\description{
COnstruct and calibrate a phylogenetic tree from sequence data
}
