% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prevalence.calculator.R
\name{prevalence.calculator}
\alias{prevalence.calculator}
\title{Calculate HIV prevalence, overall and stratified.}
\usage{
prevalence.calculator(datalist = datalist, agegroup = c(15, 30),
  timepoint = 30)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata()}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound) that should be retained, e.g. c(15, 30)}

\item{timepoint}{Point in time at which the HIV prevalence should be calculated.}
}
\value{
a dataframe with prevalence estimate and surrounding confidence bounds,
for the specified time point and age group, overall, and stratified by gender
}
\description{
Calculate the HIV prevalence at a point in time, for specific age groups and gender strata.
}
\examples{
prevalence.df <- prevalence.calculator(datalist = datalist, agegroup = c(15, 30), timepoint = 30)
}

