% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readthedata.R
\name{readthedata}
\alias{readthedata}
\title{Read the Simpact output files.}
\usage{
readthedata(modeloutput)
}
\arguments{
\item{modeloutput}{The object that is produced by RSimpactCyan::simpact.run.}
}
\value{
A list, containing dataframes for the output of the model run:
ptable (people), rtable (relationships), etable (events), ttable (HIV treatment episodes),
itable (input parameters), ltable (bookkeeping log) if created and vltable (HIV viral load)
}
\description{
Read the .csv files and combine them into one list.
}
