% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readthedata.R
\name{readthedata}
\alias{readthedata}
\title{Read the Simpact output files.}
\usage{
readthedata(modeloutput)
}
\arguments{
\item{modeloutput}{The object that is produced by \code{\link{RSimpactCyan::simpact.run()}}}
}
\value{
A list, containing dataframes for the output of the model run:
ptable (people), rtable (relationships), etable (events), ttable (HIV treatment episodes),
itable (input parameters), and ltable (bookkeeping log) if created
}
\description{
Read the .csv files and combine them into one list.
}
\examples{
cfg <- initiate()
modeloutput <- simpact.run(configParams = cfg, destDir = "~TMP")
datalist <- readthedata(modeloutput)
}

