% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.rate.calculator.R
\name{relationship.rate.calculator}
\alias{relationship.rate.calculator}
\title{Relationship rate}
\usage{
relationship.rate.calculator(datalist = datalist, timewindow = c(20, 40),
  int = FALSE, by = 1)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound)
that should be retained, e.g. c(20, 40)}

\item{int}{Logic parameter which is FALSE by deafault and means tha only one value for the relationship rate
needed, it is TRUE when the time window is being split in small time interval and get different
relationship rates within these time intervals}

\item{by}{Time step unit to subdivide the time window

data(datalist)
rels.rate <- relationship.rate.calculator(datalist = datalist,
timewindow = c(20, 40), int = FALSE, by=1)
rels.rate}
}
\description{
Calculate the relationship rate of the entire population, averaged over a particular time window.
}
