% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.fit.R
\name{sampler.fit}
\alias{sampler.fit}
\title{adjusted mice sampler}
\usage{
sampler.fit(p, data, where, m, imp, r, visitSequence, fromto, printFlag, ...)
}
\arguments{
\item{p}{padded object}

\item{data}{dataframe with missing data in it}

\item{where}{points to missing data}

\item{m}{number of cycles I think}

\item{imp}{number of imputations}

\item{r}{Forgot}

\item{visitSequence}{Default is 1 to ncol(data)}

\item{fromto}{Forgot}

\item{printFlag}{Want a flag}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Augmented to also return details of the regression model and residual error sampled.
}
