% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespentsingle.calculator.R
\name{timespentsingle.calculator}
\alias{timespentsingle.calculator}
\title{Calculate time spent being single.}
\usage{
timespentsingle.calculator(datalist = datalist, agegroup = c(15, 30),
  timewindow = c(20, 30), type = "Strict")
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{agegroup}{Boundaries of the age group (lower bound <= age < upper bound)
that should be retained, e.g. c(15, 30)}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound)
that should be retained, e.g. c(20, 30)}

\item{type}{If type is "Strict", all time spent being not being in any relationship
will be excluded from exposure time.
If type is "Harling", time will be excluded from exposure time in blocks of one year,
if the person spent that entire block not in any relationship.}
}
\value{
a vector with time spent not being in any relationships, per person.
}
\description{
Calculate the time spent not being in any relationships.
}
\examples{
data(datalist)
timespentsingle <- timespentsingle.calculator(datalist = datalist,
agegroup = c(15, 30), timewindow = c(20, 30), type = "Strict")
timespentsingle

}
