% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.network2tree.R
\name{trans.network2tree}
\alias{trans.network2tree}
\title{Build a transmission tree}
\usage{
trans.network2tree(transnetwork = transnetwork)
}
\arguments{
\item{transnetwork}{The phylo object produced by \code{\link{transmission.network.builder}}.}
}
\value{
Transmission tree object (a phylo object)
}
\description{
Build a transmission tree
}
\examples{
transm.ls <- transmission.network.builder(datalist = datalist, endpoint = 40)
transnetwork <- transm.ls[[16]]
tree0 <- trans.network2tree(transnetwork = transnetwork) # transmission tree of seeds 16

}
