% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.network2tree.R
\name{trans.network2tree}
\alias{trans.network2tree}
\title{Build a transmission tree}
\usage{
trans.network2tree(transnetwork = transnetwork)
}
\arguments{
\item{transnetwork}{The phylo object produced by \code{\link{transmNetworkBuilder.baseline}} or \code{\link{transmNetworkBuilder.diff()}}.}
}
\value{
Transmission tree object (a phylo object)
}
\description{
From the output of the \code{\link{transmNetworkBuilder.baseline()}} or \code{\link{transmNetworkBuilder.diff()}},
we can build the tranmission tree
}
\examples{

transm.ls <- transmNetworkBuilder.baseline(datalist = datalist, endpoint = 40)
transnetwork <- transm.ls[[16]]
tree0 <- trans.network2tree(transnetwork) # transmission tree of seeds 16

}
