% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.rate.calculator.R
\name{transmission.rate.calculator}
\alias{transmission.rate.calculator}
\title{Calculate the transmission rate}
\usage{
transmission.rate.calculator(datalist = datalist, timewindow = c(20, 40),
  int = FALSE, by = 1)
}
\arguments{
\item{datalist}{The datalist that is produced by \code{\link{readthedata}}}

\item{timewindow}{Boundaries of the time window (lower bound < time <= upper bound)
that should be retained, e.g. c(20, 40)}

\item{int}{Logic parameter which is FALSE by deafault meaning that only one value for the transmission rate
will be computed, it is TRUE when the time window is being split in small time interval and get different
transmission rates within these time intervals}

\item{by}{Time step unit to subdivide the time window
data(datalist)
transm.rate <- transmission.rate.calculator(datalist = datalist,
timewindow = c(20, 40), int = FALSE, by=1)
transm.rate}
}
\description{
Calculate the transmission rate of the entire population, averaged over a particular time window.
}
